Add helper method for iterating over all clusters
Iterating over all clusters will be necessary in multiple places, but was quite a pain. Here, I introduced a method that gets a lambda and executes that for all clusters. The VerletClusterTraversal also uses it already.

Because of that, a lot of parameters were no more necessary for the traverseParticlePairs() method in the VerletClustersTraversalInterface. Instead, the container is passed to enable the traversals to call the new helper method.