Add newton 3 neighbor list to VerletClusterLists
Also added test that checks it a little for validity.
Renamed AosToSoaMap to ClusterIndexMap, because it really just gives all clusters an index. It is now also used for the newton 3 neighbor list building.
The newton 3 neighbor list only contains neighbors that have a higher or equal index than itself.