Fix data race in VerletClustersColoringTraversal
The problem was that the indexing of the clusters wasn't correct when multiple grids were in one color cell. Correct here means that the newton 3 neighbor list can contain only the neighbor clusters with a higher index (or itself) and there will be no data races. Some doc is there for this on VerletClusterLists::buildClusterIndexMap().