Change signature of SoAFunctor() and add SoA to ClusterColoringTraversal
The new SoA version of VerletClustersColoringTraversal uses one SoA per grid, so not one SoA per cluster as the other traversal, because a grid is essentially a particle cell which can be easily used to load and extract the particles. To be able to let the clusters interact with each other, the signatures of the SoAFunctor() that take one or two SoAs were changed to let the user determine which particles of each SoA should be considered. The implementation changes were minimal.