Change SoAFunctor() signatures to use SoAView instead of SoA&
It was kind of hard to decide whether to use SoAView, const SoAView&, or SoAView&& as type in the signatures. SoAView& fell out pretty early, because it cannot accept rvalues, which is needed by nearly every call to SoAFunctor const SoAView& would only be possible if SoAView::begin was const, too. It would be possible to implement that, but this prevents having "real" const SoAViews which can only be read and not written to. SoAView&& is the most similar to SoAView and would also be a good option, I think. The problem here is, that if a SoAView was an lvalue, it would need to be moved to pass it to SoAFunctor() which seems unnecessary. The option that I chose, SoAView, is the most broadly applicable option and , I think, compilers don't copy rvalue SoAViews that are passed, but instead construct them directly inplace, so I don't expect much performance impact.