Fix data race in VerletClustersColoringTraversal
The color cells had the wrong size, only at least the cutoff, not the interaction length. AoS has worked without problems probably because there is no particle movement in the tests, and only some of the SoA Functors write force zero even to particles they are not supposed to work on because of the interaction length. This should maybe be caught also at some time.