Remove public getTowers() method in VerletClusterLists
This is a bit tricky, since it was the only method that gave traversals direct access to the clusters. The other traversals besides the
verlet-clusters-static all didn't need that. They could just use iteration methods provided by the VerletClusterLists class. My solution to this is
analog to the LinkedCells container. Just give the towers to the VerletClustersTraversalInterface separately and let the traversals
access it that way. This makes sure that only traversals of the container can access the towers (and thus the clusters) freely.