split off load estimator into new option
When creating new load estimation heuristics, it is no longer
necessary to add a new traversal option. Instead the new algorithm has
to be registered in autopas/options/LoadEstimatorOption.h and
compatibility with different containers and traversals can be declared
in autopas/containers/loadEstimatos.h. New algorithms should also be
implemented there and then called in
<containerType>::getLoadEstimatorFunction() to add support for the containertype.