getParticle() method for hierarchicalGrids
This is required for iterating over the particles in the container and is ultimately supposed to
prevent the particle copying.

The method relies on encoding the level of the grid and the cell index
into a combined value, which can also be decoded again. Other than that,
the method relies on functionality provided by the Linked Cells.

If we iterate over the Hierarchical Grids and we are in the valid part of
a given level, we simply return a pointer to a particle and cell and particle indices.

If a nullpointer is obtained from applying getParticle to the current HG-level
we can have two outcomes:
First:  The end of the valid part of the current level is reached but
        the current level is not the smallest level of the Hierarchical Grid.
        This means we have to advance to the next smaller level and continue there.
Second: The end of the valid part of the Hierarchical Grid is reached and there is
        nothing left.