Removed cross-level interaction by crossLevel
Removed the iteratePairwise between cross levels by generating a new LC-container
and copying particles of two levels into this container, applying iteratePairwise()
to it and then copying the particles back to their original containers.

The new method relies on getting the regionIterators of smaller particles
around each particle of the larger HG-level and directly applying the DEMFunctor
while iterating.