Fix compiling with AUTOPAS_LOG_ITERATIONS=ON
In the IterationLogger.cpp there is a call to an info function from
spdlog that it is only triggered when compiled with AUTOPAS_LOG_ITERATIONS=ON.
This call in turn was triggering a compile error of the form:

```
call to consteval function 'fmt::basic_format_string<char, std::string>::basic_format_string<std::string, 0>'
is not a constant expression
```

As described in this issue: https://github.com/fmtlib/fmt/issues/2438, the
solution is to surround the parameter in a fmt::runtime, such that we
can compile it in a non-consteval function.

I am not sure whether this is handled better in newer versions of spdlog and if so,
it might be a good idea to also update it.