Remove redundant template parameter in default constructor
I have been doing some internal checks with GCC11 and the code
`ParticleVector<Particle_T>() = default;` should not have compiled in the first place,
as one gets an `error: expected unqualified-id before ')' token`. Below is a
reproducible example that I tested in compiler explorer with GCC-11.4:

```
template<class T>
class A {
public:
    A<T>() = default;
private:
    T _a;
};

int main()
{
    A<int> a;
}
```

The above code would work only with GCC-14 and newer, and would anyway generate
a warning. I am not sure why the code was compiling before, as this should be
unrelated with the C++20 migration.