Update the ALL library to v0.9.3 to fix a compiling error
I was having trouble compiling the previous version, as I was getting the
following compilation error:

```
/__w/AutoPas/AutoPas/build/_deps/allfetch-src/src/../include/ALL_Point.hpp:116:3: error: template-id not allowed for destructor
  116 |   ~Point<T>(){};
 ```
As before, the easy fix was to remove the template argument from the destructor.
Nonetheless, since this is an external dependency, I considered a better idea to
simply upgrade the library to a newer version that had fixed that error.