Reimplement fast particle buffer and add new CMake option
This commit reintroduces the fast particle buffer and adds a dedicated CMake option for enabling it.
The feature is currently not functional when AUTOPAS_ENABLE_DYNAMIC_CONTAINERS=OFF and AUTOPAS_ENABLE_FAST_PARTICLE_BUFFER=ON. Some dynamic container features are still missing in fast buffer mode, if the dynamic container is disabled. Add logging for particle buffer size, number of owned particles, number of halo particles, and number of fast particles.